requirements.txt
esptool>=4.0
pyserial>=3.0
